34: journal_accounts: [select2 untuk edit akun, search akunnya], fix: accountsp: [delete nya ngga bisa kalau ad transaksi], bug: variables: [accounting module always on]

+++ app/Http/Controllers/Primary/Inventory/AccountController.php
@@ -206,7 +206,16 @@ class AccountController extends Controller
 
     public function destroy(Request $request, String $id)
     {
-        $account = Inventory::findOrFail($id);
+        $account = Inventory::with('tx_details')->findOrFail($id);
+
+        // check apakah ada transaksi
+        if ($account->tx_details->count() > 0) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Account has related transaction. Cannot delete.',
+            ], 500);
+        }
+
 
         $account->delete();
